some refactors of the pkg/test code
