+++ test/gke/location.go
