+++ test/gke/location_test.go
