+++ testutils/clustermanager/gke.go
@@ -23,9 +23,9 @@ import (
 	"strings"
 
 	container "google.golang.org/api/container/v1beta1"
+	"knative.dev/pkg/test/gke"
 	"knative.dev/pkg/testutils/clustermanager/boskos"
 	"knative.dev/pkg/testutils/common"
-	"knative.dev/pkg/testutils/gke"
 )
 
 const (
